some packages refactoring
