--- src/java/homework3/Client.java
@@ -1,4 +1,0 @@
-package java.homework3;
-
-public interface Client {
-}
