--- src/java/homework3/Contract.java
@@ -1,4 +1,0 @@
-package java.homework3;
-
-public interface Contract {
-}
