--- src/java/homework3/ContractFactoryMethod.java
@@ -1,5 +1,0 @@
-package java.homework3;
-
-public interface ContractFactoryMethod {
-    Contract createContract(Manager manager, Client client);
-}
