+++ src/java/homework3/client/Client.java
@@ -1,0 +1,4 @@
+package java.homework3.client;
+
+public interface Client {
+}
