+++ src/java/homework3/client/ClientImpl.java
@@ -1,4 +1,4 @@
-package java.homework3;
+package java.homework3.client;
 
 public class ClientImpl implements ConsumerClient, ProviderClient {
 }
