+++ src/java/homework3/client/ProviderClient.java
@@ -1,4 +1,4 @@
-package java.homework3;
+package java.homework3.client;
 
 public interface ProviderClient extends Client {
 }
