+++ src/java/homework3/contract/Contract.java
@@ -1,0 +1,4 @@
+package java.homework3.contract;
+
+public interface Contract {
+}
