+++ src/java/homework3/contract/ProviderContract.java
@@ -1,4 +1,4 @@
-package java.homework3;
+package java.homework3.contract;
 
 public interface ProviderContract extends Contract {
 }
